+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.4.0](https://github.com/mrazauskas/tsd-lite/compare/v0.3.0...v0.4.0) (2022-01-17)
+
 ### ⚠ BREAKING CHANGES
 
 - do not add flattened `message` to `tsdResults` object. `messageText` is there already and should be used instead.